Expose initial item for setup types
