+++ library/src/main/java/com/efedaniel/storytablayout/setup/ViewPagerSetupType.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.efedaniel.storytablayout.setup.setuptype
+package com.efedaniel.storytablayout.setup
 
 import androidx.viewpager.widget.ViewPager
 import androidx.viewpager.widget.ViewPager.OnPageChangeListener
@@ -30,6 +30,10 @@ internal class ViewPagerSetupType(private val viewPager: ViewPager) : SetupType,
 
     override fun onCurrentBarFilled(nextIndex: Int) {
         viewPager.setCurrentItem(nextIndex, true)
+    }
+
+    override fun getInitialPage(): Int {
+        return viewPager.currentItem
     }
 
     override fun onPageScrolled(position: Int, positionOffset: Float, positionOffsetPixels: Int) {
